refonte

+++ src/api/apropos.js
@@ -22,20 +22,24 @@ module.exports = (app) => {
   			});
   		});
   		ap.post("/apropos/modifApropos",(req,res)=>{
-  			const apropos = req.body.u;
-  			const q= "DELETE FROM apropos";
-  			connection.query(q,(e,r,f)=>{
-  				if (e) throw e;
-  				else{
-  					const q2 = "INSERT INTO apropos (apropos) VALUE ('"+apropos+"')";
-  					connection.query(q2, (e2,r2,f2)=>{
-  						if (e2) throw e2;
-  						else{
-  							res.send("OK");
-  						}
-  					});
-  				}
-  			});
+        if(req.session && req.session.adminConnected){
+    			const apropos = req.body.u;
+    			const q= "DELETE FROM apropos";
+    			connection.query(q,(e,r,f)=>{
+    				if (e) throw e;
+    				else{
+    					const q2 = "INSERT INTO apropos (apropos) VALUE ('"+apropos+"')";
+    					connection.query(q2, (e2,r2,f2)=>{
+    						if (e2) throw e2;
+    						else{
+    							res.send("OK");
+    						}
+    					});
+    				}
+    			});
+        }else{
+          res.send("ER");
+        }
   		});
   	};
   	return aproposAPI(app);
